chore(connd): re-org modules
