+++ orb-connd/src/cellular/mod.rs
@@ -1,0 +1,6 @@
+use color_eyre::Result;
+use tokio::task::{self, JoinHandle};
+
+pub async fn ensure_connectivity() -> JoinHandle<Result<()>> {
+    task::spawn(async move { Ok::<_, color_eyre::Report>(()) })
+}
