+++ orb-connd/src/lib.rs
@@ -1,19 +1,20 @@
 use color_eyre::eyre::{Result, WrapErr};
-use modem::{modem_manager, net_stats::NetStats, Modem};
 use orb_info::orb_os_release::{OrbOsPlatform, OrbOsRelease};
 use std::time::Duration;
 use tokio::{
     fs,
     signal::unix::{self, SignalKind},
+    task::JoinHandle,
 };
-use tracing::{error, info, warn};
-use utils::{retry_for, State};
+use tracing::{info, warn};
+use utils::retry_for;
 
-mod backend_status_reporter;
-mod dd_reporter;
-mod modem;
-mod modem_monitor;
+mod cellular;
+mod modem_manager;
+mod telemetry;
 mod utils;
+
+pub type Tasks = Vec<JoinHandle<Result<()>>>;
 
 pub async fn run() -> Result<()> {
     // TODO: this is temporary while this daemon only supports cellular metrics
@@ -35,19 +36,8 @@ pub async fn run() -> Result<()> {
         return Ok(());
     }
 
-    info!("getting initial modem information");
-    let modem = retry_for(
-        Duration::from_secs(120),
-        Duration::from_secs(10),
-        make_modem,
-    )
-    .await?;
-
-    let modem_monitor_handle =
-        modem_monitor::start(modem.clone(), Duration::from_secs(20));
-    let backend_status_reporter_handle =
-        backend_status_reporter::start(modem.clone(), Duration::from_secs(30));
-    let dd_reporter_handle = dd_reporter::start(modem, Duration::from_secs(20));
+    let mut tasks = vec![cellular::ensure_connectivity().await];
+    tasks.extend(telemetry::start().await?);
 
     let mut sigterm = unix::signal(SignalKind::terminate())?;
     let mut sigint = unix::signal(SignalKind::interrupt())?;
@@ -58,28 +48,12 @@ pub async fn run() -> Result<()> {
     }
 
     info!("aborting tasks and exiting gracefully");
-    modem_monitor_handle.abort();
-    backend_status_reporter_handle.abort();
-    dd_reporter_handle.abort();
+
+    for handle in tasks {
+        handle.abort();
+    }
 
     Ok(())
-}
-
-async fn make_modem() -> Result<State<Modem>> {
-    let modem: Result<Modem> = async {
-        let modem_id = modem_manager::get_modem_id().await?;
-        let imei = modem_manager::get_imei(&modem_id).await?;
-        let iccid = modem_manager::get_iccid().await?;
-        let state = modem_manager::get_connection_state(&modem_id).await?;
-        modem_manager::start_signal_refresh(&modem_id).await?;
-        let net_stats = NetStats::from_wwan0().await?;
-
-        Ok(Modem::new(modem_id, iccid, imei, state, net_stats))
-    }
-    .await
-    .inspect_err(|e| error!("make_modem: {e}"));
-
-    Ok(State::new(modem?))
 }
 
 async fn wwan0_exists() -> Result<bool> {
