+++ orb-connd/src/main.rs
@@ -2,6 +2,12 @@ use color_eyre::eyre::Result;
 
 const SYSLOG_IDENTIFIER: &str = "worldcoin-connd";
 
+// - [ ] separate telemetry into telemetry
+// - [ ] determine config
+// - [ ] modem manager logic
+//      - [ ] further breakdown
+// - [ ] test with wifi qr code on dev orb, from boot
+
 #[tokio::main]
 async fn main() -> Result<()> {
     color_eyre::install()?;
