+++ orb-connd/src/utils.rs
@@ -5,6 +5,7 @@ use std::{
 };
 use tokio::{
     process::Command,
+    task::JoinHandle,
     time::{self, Instant},
 };
